extract intcode
