+++ aoc2019/aoc2019/day_02.py
@@ -1,0 +1,8 @@
+from .intcode import IntCode
+
+
+def part1():
+    intcode = IntCode()
+    intcode.get_input("input/day_02")
+    value = intcode.diagnostic_program(None)
+    print(value)
