+++ aoc2019/input/day_02_ex1
