+++ aoc2019/pyproject.toml
@@ -5,6 +5,7 @@ description = ""
 authors = ["Steven Powell <princeodd@gmail.com>"]
 
 [tool.poetry.scripts]
+d2p1 = "aoc2019.day_02:part1"
 d5p1 = "aoc2019.day_05:part1"
 d5p2 = "aoc2019.day_05:part2"
