--- aoc2019/tests/data/day_05_part2_example1
@@ -1,1 +1,0 @@
-3,12,6,12,15,1,13,14,13,4,13,99,-1,0,1,9
