--- aoc2019/tests/data/day_05_part2_example2
@@ -1,1 +1,0 @@
-3,3,1105,-1,9,1101,0,0,12,4,12,99,1
