--- aoc2019/tests/test_aoc2019.py
@@ -1,5 +1,0 @@
-from aoc2019 import __version__
-
-
-def test_version():
-    assert __version__ == '0.1.0'
